Add RichestCustomer method and unit tests
Implemented the `RichestCustomer` method in the `MatrixProblems` class within the `LeetcodeMsft.Lib` namespace. This method calculates the wealth of the richest customer based on their accounts and throws an `ArgumentNullException` if the input is null.

Added unit tests for the `RichestCustomer` method in a new test class `MatrixProblemTests` using FluentAssertions for assertions. Introduced a new namespace `LeetcodeMsft.Lib.Tests` for the test class. Test cases include verifying correct functionality and ensuring an exception is thrown for null input.